Split out URL utilities into their own file and use them
